Improve und fix various things
BlockActions: Use unsigned byte (short)
BlockActions: Add BellAction and ignore all other actions
ChatComponent: Fix mishandling, wrong order
Rename some GameStates
Fix Respawn Packet Handling in 1.16.2+

+++ src/main/java/de/bixilon/minosoft/data/mappings/blocks/actions/EndGatewayAction.java
@@ -15,7 +15,7 @@ package de.bixilon.minosoft.data.mappings.blocks.actions;
 
 public class EndGatewayAction implements BlockAction {
 
-    public EndGatewayAction(byte status, byte ignored) {
+    public EndGatewayAction(short status, short ignored) {
         // only 1 action (id 1)
     }
